añadida autenticación jwt

+++ server.js
@@ -1,6 +1,4 @@
 //SETUP//
-
-
 let express = require('express')
 require('dotenv').config()
 let app = express()
@@ -16,11 +14,24 @@ app.listen(port, () => {
 let bodyParser = require('body-parser')
 midware = bodyParser.urlencoded({extended: false})
 app.use(bodyParser.json());
+//BCRYPT
+const bcrypt = require('bcrypt');
 
 //DATABASE//
 const mongoose = require('mongoose')
 const {Schema} = mongoose;
 mongoose.connect(process.env.MONGO_URI)
+
+//AUTH//
+const jwt = require("jsonwebtoken")
+
+
+
+
+
+
+
+/////////////////////////BLOG/////////////////////////
 const paragraphSchema = new Schema({
     title: String, // String is shorthand for {type: String}
     author: String,
@@ -29,7 +40,7 @@ const paragraphSchema = new Schema({
 });
 const Paragraph = mongoose.model('paragraphs', paragraphSchema);
 
-const getMaxId = async () => {
+const getMaxId = async function () {
     try {
         const maxIdParagraph = await Paragraph.findOne({}, null, {sort: {id: -1}})
         return maxIdParagraph.id
@@ -39,12 +50,11 @@ const getMaxId = async () => {
     }
 };
 
-
-//FUNCIONES//
-
-
 addParagraph = async function (req, res) {
     let id
+    console.log(req.body.title)
+    console.log(req.user.username)
+    console.log(req.body.content)
     await getMaxId()
         .then((maxId) => {
             id = maxId+1
@@ -53,9 +63,10 @@ addParagraph = async function (req, res) {
             console.error(error)
             return res.status(500).send({error: error.message})
         })
+
     let parrafoNuevo = new Paragraph({
         title: req.body.title,
-        author: req.body.author,
+        author: req.user.username,
         content: req.body.content,
         id: id
     })
@@ -67,12 +78,7 @@ addParagraph = async function (req, res) {
         })
 
 }
-
-
-//ENDPOINTS//
-
-
-app.get('/fetch', async (req, res) => {
+app.get('/fetch',authenticateToken, async (req, res) => {
     await Paragraph.find()
         .then( docs => {
             let array = []
@@ -87,15 +93,101 @@ app.get('/fetch', async (req, res) => {
         })
 })
 
-app.get('/', (req, res) => {
+app.get('/', authenticateToken, (req, res) => {
+    console.log(req.user)
     res.send('hello world')
 })
 
-app.post('/push', midware, addParagraph)
+app.post('/push', midware, authenticateToken, addParagraph)
 
-app.delete('/remove', async (req, res) => {
+app.delete('/remove', authenticateToken, async (req, res) => {
     const received_id = parseInt(req.query.id)
-    const result = await Paragraph.deleteMany({id:received_id})
+    const result = await Paragraph.deleteMany({id:received_id, author:req.user.username})
     res.sendStatus(200)
 })
 
+
+
+
+
+
+
+
+/////////////////////////LOGIN/////////////////////////
+
+const credenciales_validas = async function(username, password){
+    const foundUser = await User.findOne({username:username})
+    if (foundUser.hash==null) {
+        return false
+    }
+    return bcrypt.compareSync(password, foundUser.hash);
+    
+}
+
+function authenticateToken(req, res, next){
+    const authHeader = req.headers['authorization']
+    const token = authHeader && authHeader.split(' ')[1]
+    if (token==null){
+        return res.sendStatus(401)
+    }
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err,user) =>{
+        if (err) return res.sendStatus(403)
+        req.user = user
+        next()
+    })
+}
+
+
+const userSchema = new Schema({
+    email: String,
+    username: String,
+    hash: String,
+})
+const User = mongoose.model('users', userSchema);
+
+const isUsernameInUse = async function(username){
+    if(await User.findOne({username:username})!=null){
+        return true
+    }
+    return false
+}
+const isEmailInUse = async function(email){
+    if(await User.findOne({email:email})!=null){
+        return true
+    }
+    return false
+}
+
+const addNewUser = async function (req, res) {
+    if(await isUsernameInUse(req.body.username)){
+        return res.status(409).send({error: "Username ya en uso"})
+    }
+    if(await isEmailInUse(req.body.username)){
+        return res.status(409).send({error: "Email ya en uso"})
+    }
+    const salt = bcrypt.genSaltSync(10);
+    const hash = bcrypt.hashSync(req.body.password, salt);
+
+    let usernNuevo = new User({
+        email: req.body.email,
+        username: req.body.username,
+        hash: hash,
+    })
+    await usernNuevo.save()
+        .then( () => {
+            res.sendStatus(200)
+         })
+        .catch((error) => {
+            console.log(error)
+        })
+}
+app.post('/signupform', midware, addNewUser)
+
+
+app.post('/login', async (req, res) =>{
+    if(await credenciales_validas(req.body.username, req.body.password)==false){
+        return res.sendStatus(403)
+    }
+    const accessToken = jwt.sign({username: req.body.username, password: req.body.password}, process.env.ACCESS_TOKEN_SECRET, { expiresIn: 20000 })
+    res.json({accessToken: accessToken})
+})
